% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{form}
\alias{form}
\title{FORM Algorithm using Rcpp}
\usage{
form(g, mean, sigma)
}
\arguments{
\item{g}{A function representing the limit state function}

\item{mean}{A NumericVector of the mean values of the input variables}

\item{sigma}{A NumericVector of the standard deviations of the input variables}
}
\value{
A List with failure probability \code{pf} and reliability index \code{beta}
}
\description{
Calculates the failure probability and reliability index using the FORM algorithm
}
\examples{
\dontrun{
g <- function(x) { x[1] + x[2] - 3 }
mean <- c(2, 2)
sigma <- c(1, 1)
result <- form(g, mean, sigma)
print(result)
}
}
