% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_functions.R
\name{predict_svm}
\alias{predict_svm}
\title{Predict Failure Probability and Reliability Index Based on the Trained SVM Model}
\usage{
predict_svm(svm_model, data_predict, g)
}
\arguments{
\item{svm_model}{A trained SVM model (object of class \code{svm}).}

\item{data_predict}{A matrix or data frame of samples to be predicted.}

\item{g}{A function defining the limit state function, which takes a matrix.}
}
\value{
A list containing:
\itemize{
\item Pf: Estimated failure probability (proportion of failure events).
\item beta: Reliability index (distance to the limit state surface).
}
}
\description{
This function uses a trained SVM model to predict the failure probability and reliability index
based on the provided and limit state function.
}
\examples{
library(riar)
set.seed(123)
g <- function(x) {
  return(x[1]^2 + x[2]^2 - 3)
}
n_samples <- 500  # Define the number of samples
x <- matrix(rnorm(2*n_samples), ncol = 2)
svm_model <- train_svm(x, g)  # Train the SVM model
results <- predict_svm(svm_model, x, g)  # Predict failure probability and reliability index
print(results)  # Print the results
}
