% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm_functions.R
\name{train_svm}
\alias{train_svm}
\title{SVM Model Training and Prediction}
\usage{
train_svm(x, g)
}
\arguments{
\item{x}{The input data (matrix or data frame).}

\item{g}{The limit state function.}
}
\value{
The SVM modeling result.
}
\description{
This function trains a Support Vector Machine (SVM) model on the provided data.
}
\examples{
library(riar)
set.seed(123)
g <- function(x) {
  return(x[1]^2 + x[2]^2 - 3)
}
n_samples <- 500  # Define the number of samples
x <- matrix(rnorm(n_samples*2), ncol = 2) # Generate random number
svm_model <- train_svm(x, g)  # Train the SVM model
print("The svm model has been successfully constructed")  # Print the results
}
